test(file-transfer): split directory listing policy coverage
